1. Pytorch profiler - cpu time, gpu time, memory usage, investigating modules.
This recipe explains how to use PyTorch profiler and measure the time and memory consumption of the model's operators.

Introduction - 
PyTorch includes a simple profiler API that is useful when user needs to determine the most expensive operators in the model.
In this recipe, we will use a simple Resnet model to demonstrate how to use profiler to analyze model performance.